+++ src/main/java/store/lastdance/controller/expense/ExpenseController.java
@@ -209,5 +209,43 @@ public class ExpenseController {
         return ResponseEntity.ok(ApiResponse.success(response));
     }
 
-    
+    @PostMapping("/analyze")
+    @RateLimit // 30초에 1번만 요청 가능
+    @Operation(summary = "LLM 지출 분석 요청", description = "지정된 기간의 지출 내역을 LLM을 통해 분석")
+    public ResponseEntity<ApiResponse<AnalyzeExpenseResponseDTO>> analyzeExpenses(
+            @Parameter(hidden = true) @AuthenticationPrincipal CustomOAuth2User oAuth2User,
+            @Valid @RequestBody AnalyzeExpenseRequestDTO requestDTO
+    ) {
+        UUID userId = oAuth2User.getUserId();
+        AnalyzeExpenseResponseDTO response = expenseService.analyzeExpenses(userId, requestDTO);
+        return ResponseEntity.ok(ApiResponse.success(response));
+    }
+
+    @PostMapping("/analyze/{historyId}/feedback")
+    @Operation(summary = "LLM 지출 분석 피드백", description = "LLM 지출분석 결과에 대해 피드백(좋아요/싫어요) 를 남깁니다.")
+    public ResponseEntity<ApiResponse<String>> feedbackAnalyzeExpense(
+            @Parameter(hidden = true) @AuthenticationPrincipal CustomOAuth2User oAuth2User,
+            @PathVariable Long historyId,
+            @RequestParam String type
+    ) {
+        UUID userId = oAuth2User.getUserId();
+        String result = expenseService.toggleFeedback(historyId, userId, type);
+        return ResponseEntity.ok(ApiResponse.success(result));
+    }
+
+
+    @GetMapping("/analyze/history")
+    @Operation(summary = "LLM 지출 분석 내역 조회", description = "사용자의 전체 지출 분석 내역을 최신순으로 조회 (페이징 포함)")
+    public ResponseEntity<ApiResponse<PageWithSummaryResponse<ExpenseAnalysisHistoryDTO>>> getAnalysisHistoryList(
+            @Parameter(hidden = true) @AuthenticationPrincipal CustomOAuth2User oAuth2User,
+            @PageableDefault(
+                    sort = "createdAt",
+                    direction = Sort.Direction.DESC
+            ) Pageable pageable
+    ) {
+        UUID userId = oAuth2User.getUserId();
+        PageWithSummaryResponse<ExpenseAnalysisHistoryDTO> response = expenseService.getExpenseAnalysisHistory(userId, pageable);
+
+        return ResponseEntity.ok(ApiResponse.success(response));
+    }
 }
