+++ src/main/java/store/lastdance/dto/expense/AnalyzeExpenseResponseDTO.java
@@ -1,6 +1,7 @@
-package store.lastdance.dto.analysis;
+package store.lastdance.dto.expense;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.math.BigDecimal;
@@ -84,3 +85,4 @@ public record AnalyzeExpenseResponseDTO(
         int transactionCount
     ) {}
 }
+
