+++ src/main/java/store/lastdance/dto/expense/ExpenseAnalysisHistoryDTO.java
@@ -1,8 +1,8 @@
-package store.lastdance.dto.analysis;
+package store.lastdance.dto.expense;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import io.swagger.v3.oas.annotations.media.Schema;
-import store.lastdance.domain.analysis.ExpenseAnalysisHistory;
+import store.lastdance.domain.expense.ExpenseAnalysisHistory;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
