+++ src/main/java/store/lastdance/dto/expense/SaveAnalysisResultRequestDTO.java
@@ -2,8 +2,6 @@ package store.lastdance.dto.expense;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
-import store.lastdance.dto.analysis.AnalyzeExpenseRequestDTO;
-import store.lastdance.dto.analysis.AnalyzeExpenseResponseDTO;
 
 public record SaveAnalysisResultRequestDTO(
         @NotNull @Valid AnalyzeExpenseRequestDTO requestDTO,
