+++ src/main/java/store/lastdance/repository/expense/ExpenseAnalysisHistoryRepository.java
@@ -1,4 +1,4 @@
-package store.lastdance.repository.analysis;
+package store.lastdance.repository.expense;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -8,13 +8,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import store.lastdance.domain.analysis.ExpenseAnalysisHistory;
+import store.lastdance.domain.expense.ExpenseAnalysisHistory;
 import store.lastdance.domain.user.User;
 import store.lastdance.dto.admin.AdminExpenseAnalyzerHistoryDTO;
 import store.lastdance.dto.admin.ExpenseAnalyzerFeedbackStatsDTO;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.UUID;
 
 public interface ExpenseAnalysisHistoryRepository extends JpaRepository<ExpenseAnalysisHistory, Long>,
         JpaSpecificationExecutor<ExpenseAnalysisHistory> {
@@ -22,12 +23,14 @@ public interface ExpenseAnalysisHistoryRepository extends JpaRepository<ExpenseA
     @EntityGraph(attributePaths = "user")
     Page<ExpenseAnalysisHistory> findAll(Specification<ExpenseAnalysisHistory> spec, Pageable pageable);
 
-    
+    @Query("SELECT new store.lastdance.dto.admin.AdminExpenseAnalyzerHistoryDTO(" +
+           "eah.id, eah.user.email, eah.user.nickname, eah.createdAt, eah.up, eah.down) " +
+           "FROM ExpenseAnalysisHistory eah JOIN eah.user u")
+    Page<AdminExpenseAnalyzerHistoryDTO> findHistoryProjection(Specification<ExpenseAnalysisHistory> spec, Pageable pageable);
 
     // 특정 사용자의 모든 분석 내역을 최신순으로 조회
     List<ExpenseAnalysisHistory> findByUserOrderByCreatedAtDesc(User user);
 
-    @EntityGraph(attributePaths = "user")
     Page<ExpenseAnalysisHistory> findByUser(User user, Pageable pageable);
 
     List<ExpenseAnalysisHistory> findByCreatedAtBetween(LocalDateTime startDate, LocalDateTime endDate);
